convert dasris examples to typescript
