+++ back/src/common/workflow.d.ts
@@ -6,17 +6,17 @@
  * Worflows definition is parsed by the doc project to create code examples in the documentation
  */
 export type Workflow = {
-  "Title of the workflow";
+  // Title of the workflow
   title: string;
-  "Longer description of the workflow explaining the traceability use case";
+  // Longer description of the workflow explaining the traceability use case;
   description?: string;
-  "Name and profile of the companies involved in the workflow";
+  // Name and profile of the companies involved in the workflow
   companies: { name: string; companyTypes: string[] }[];
-  "List of steps to be applied to the BSD";
+  // List of steps to be applied to the BSD
   steps: WorkflowStep[];
-  "Mocked context used in the documentation code examples";
+  // Mocked context used in the documentation code examples
   docContext: any;
-  "Optional Mermaid chart definition representing the workflow";
+  // Optional Mermaid chart definition representing the workflow
   chart?: string;
 };
 
@@ -24,18 +24,18 @@ export type Workflow = {
  * A composable and testable step used in a workflow
  */
 export type WorkflowStep = {
-  "Description of the step";
+  // Description of the step
   description: string;
-  "GraphQL mutation to apply";
+  // GraphQL mutation to apply
   mutation: string;
-  "GraphQL variables";
+  // GraphQL variables";
   variables: (ctx: any) => any;
-  "Expected result of the mutation";
+  // Expected result of the mutation
   expected: any;
-  "Name of the company performing the mutation as defined in the parent workflow companies field";
+  // Name of the company performing the mutation as defined in the parent workflow companies field
   company: string;
-  "How to parse GraphQL response to get data";
+  // How to parse GraphQL response to get data
   data: (response: any) => any;
-  "Optional callback to update the runtime context that is passed from one step to the other";
+  // Optional callback to update the runtime context that is passed from one step to the other
   setContext?: (ctx: Context, data: any) => Context;
 };
