+++ back/src/dasris/examples/workflows/index.ts
@@ -1,0 +1,5 @@
+import acheminementDirect from "./acheminementDirect";
+
+export default {
+  acheminementDirect
+};
